+++ CotEditor/Sources/CETextSelection.h
@@ -86,6 +86,7 @@ typedef NS_ENUM(NSUInteger, CEUNFType) {
 - (void)handleMoveLineDownScriptCommand:(NSScriptCommand *)command;
 - (void)handleSortLinesAscendingScriptCommand:(NSScriptCommand *)command;
 - (void)handleReverseLinesScriptCommand:(NSScriptCommand *)command;
+- (void)handleDeleteDuplicateLineScriptCommand:(NSScriptCommand *)command;
 - (void)handleCommentOutScriptCommand:(NSScriptCommand *)command;
 - (void)handleUncommentScriptCommand:(NSScriptCommand *)command;
 - (void)handleChangeCaseScriptCommand:(NSScriptCommand *)command;